requirements.txt
SQLAlchemy < 1.4
apache-airflow==2.1.0
apache-airflow-providers-http==2.0.0
apache-airflow-providers-postgres==1.0.1
apache-airflow-providers-slack==4.0.0
apache-airflow-providers-amazon==1.4.0
apache-airflow-providers-google==3.0.0
pytest==6.2.3